enable wifi now

+++ main/battery.h
@@ -8,6 +8,6 @@ void battery_start();
 void battery_stop();
 bool is_battery_level_low();
 bool is_battery_extremely_low();
-int get_battery_level();
+uint8_t get_battery_level();
 
 #endif  /*_BS_BATTERY_H_*/
